Update sfdc_log.rb
I believe this is a change in how rails 3 works with ruby 1.9, since 1.9 supports encodings it will attempt to convert all strings to whatever encoding you have set in your app configuration (application.rb), typically this is 'utf-8'. 

Many of my files are not in utf-8 and the code barfs at this point. To ensure that it works for all encoding, wouldn't it be better for this to be in binary?